# Use freelists instead of sync.Pool for value object recycling