--- a/value/list.go
+++ b/value/list.go
@@ -23,15 +23,25 @@ type List interface {
 	// At returns the item at the given position in the map. It will
 	// panic if the index is out of range.
 	At(int) Value
+	// AtUsing uses the provided allocator and returns the item at the given
+	// position in the map. It will panic if the index is out of range.
+	// The returned Value should be given back to the Allocator when no longer needed
+	// by calling Allocator.Free(Value).
+	AtUsing(Allocator, int) Value
 	// Range returns a ListRange for iterating over the items in the list.
 	Range() ListRange
-
-	// Equals compares the two list, and return true if they are the same, false otherwise.
+	// RangeUsing uses the provided allocator and returns a ListRange for
+	// iterating over the items in the list.
+	// The returned Range should be given back to the Allocator when no longer needed
+	// by calling Allocator.Free(Value).
+	RangeUsing(Allocator) ListRange
+	// Equals compares the two lists, and return true if they are the same, false otherwise.
 	// Implementations can use ListEquals as a general implementation for this methods.
 	Equals(List) bool
-	// Recycle gives back this List once it is no longer needed. The
-	// value shouldn't be used after this call.
-	Recycle()
+	// EqualsUsing uses the provided allocator and compares the two lists, and return true if
+	// they are the same, false otherwise. Implementations can use ListEqualsUsing as a general
+	// implementation for this methods.
+	EqualsUsing(Allocator, List) bool
 }
 
 // ListRange represents a single iteration across the items of a list.
@@ -41,11 +51,8 @@ type ListRange interface {
 	// Item returns the index and value of the current item in the range. or panics if there is no current item.
 	// For efficiency, Item may reuse the values returned by previous Item calls. Callers should be careful avoid holding
 	// pointers to the value returned by Item() that escape the iteration loop since they become invalid once either
-	// Item() or Recycle() is called.
+	// Item() or Allocator.Free() is called.
 	Item() (index int, value Value)
-	// Recycle gives back this ListRange once it is no longer needed. The value returned by Item() becomes invalid once this is
-	// called.
-	Recycle()
 }
 
 var EmptyRange = &emptyRange{}
@@ -60,24 +67,28 @@ func (_ *emptyRange) Item() (index int, value Value) {
 	panic("Item called on empty ListRange")
 }
 
-func (_ *emptyRange) Recycle() {}
-
 // ListEquals compares two lists lexically.
 // WARN: This is a naive implementation, calling lhs.Equals(rhs) is typically the most efficient.
 func ListEquals(lhs, rhs List) bool {
+	return ListEqualsUsing(HeapAllocator, lhs, rhs)
+}
+
+// ListEqualsUsing uses the provided allocator and compares two lists lexically.
+// WARN: This is a naive implementation, calling lhs.EqualsUsing(allocator, rhs) is typically the most efficient.
+func ListEqualsUsing(a Allocator, lhs, rhs List) bool {
 	if lhs.Length() != rhs.Length() {
 		return false
 	}
 
-	lhsRange := lhs.Range()
-	defer lhsRange.Recycle()
-	rhsRange := rhs.Range()
-	defer rhsRange.Recycle()
+	lhsRange := lhs.RangeUsing(a)
+	defer a.Free(lhsRange)
+	rhsRange := rhs.RangeUsing(a)
+	defer a.Free(rhsRange)
 
 	for lhsRange.Next() && rhsRange.Next() {
 		_, lv := lhsRange.Item()
 		_, rv := rhsRange.Item()
-		if !Equals(lv, rv) {
+		if !EqualsUsing(a, lv, rv) {
 			return false
 		}
 	}
@@ -92,10 +103,16 @@ func ListLess(lhs, rhs List) bool {
 // ListCompare compares two lists lexically. The result will be 0 if l==rhs, -1
 // if l < rhs, and +1 if l > rhs.
 func ListCompare(lhs, rhs List) int {
-	lhsRange := lhs.Range()
-	defer lhsRange.Recycle()
-	rhsRange := rhs.Range()
-	defer rhsRange.Recycle()
+	return ListCompareUsing(HeapAllocator, lhs, rhs)
+}
+
+// ListCompareUsing uses the provided allocator and compares two lists lexically. The result will be 0 if l==rhs, -1
+// if l < rhs, and +1 if l > rhs.
+func ListCompareUsing(a Allocator, lhs, rhs List) int {
+	lhsRange := lhs.RangeUsing(a)
+	defer a.Free(lhsRange)
+	rhsRange := rhs.RangeUsing(a)
+	defer a.Free(rhsRange)
 
 	for {
 		lhsOk := lhsRange.Next()
@@ -114,7 +131,7 @@ func ListCompare(lhs, rhs List) int {
 		}
 		_, lv := lhsRange.Item()
 		_, rv := rhsRange.Item()
-		if c := Compare(lv, rv); c != 0 {
+		if c := CompareUsing(a, lv, rv); c != 0 {
 			return c
 		}
 		// The items are equal; continue.