@jpbetz 2020-02-06
> This looks good, it's mostly mechanical after all. It clearly needs to be rebased/squashed though, thanks!

Thanks for looking at it so quickly. I've squashed and reduced the copy-paste in the allocator code.